remove layout folder and CallLayout, add index page, refactor CallPageLayout
