+++ src/app/call/[callAssId]/page.tsx
@@ -1,19 +1,23 @@
 import { Box } from '@mui/material';
-import { Suspense } from 'react';
 import { headers } from 'next/headers';
+import { ReactNode, Suspense } from 'react';
+import { notFound, redirect } from 'next/navigation';
 
 import BackendApiClient from 'core/api/client/BackendApiClient';
-import { ZetkinCallAssignment } from 'utils/types/zetkin';
-import AssignmentStatsPage from 'features/call/pages/AssignmentStatsPage';
+import CallPage from 'features/call/pages';
+import HomeThemeProvider from 'features/home/components/HomeThemeProvider';
+import redirectIfLoginNeeded from 'core/utils/redirectIfLoginNeeded';
 import ZUILogoLoadingIndicator from 'zui/ZUILogoLoadingIndicator';
+import { ZetkinCallAssignment } from 'utils/types/zetkin';
+import { CALL, hasFeature } from 'utils/featureFlags';
 
-interface PageProps {
-  params: {
-    callAssId: string;
-  };
-}
+type Props = {
+  children?: ReactNode;
+  params: { callAssId: string };
+};
 
-export default async function Page({ params }: PageProps) {
+const CallPageLayout = async ({ children, params }: Props) => {
+  await redirectIfLoginNeeded();
   const headersList = headers();
   const headersEntries = headersList.entries();
   const headersObject = Object.fromEntries(headersEntries);
@@ -26,24 +30,35 @@ export default async function Page({ params }: PageProps) {
   );
 
   if (!assignment) {
-    return null;
+    return notFound();
   }
+  const { callAssId } = params;
 
-  return (
-    <Suspense
-      fallback={
-        <Box
-          alignItems="center"
-          display="flex"
-          flexDirection="column"
-          height="90dvh"
-          justifyContent="center"
+  if (hasFeature(CALL, assignment.organization.id, process.env)) {
+    return (
+      <HomeThemeProvider>
+        <Suspense
+          fallback={
+            <Box
+              alignItems="center"
+              display="flex"
+              flexDirection="column"
+              height="90dvh"
+              justifyContent="center"
+            >
+              <ZUILogoLoadingIndicator />
+            </Box>
+          }
         >
-          <ZUILogoLoadingIndicator />
-        </Box>
-      }
-    >
-      <AssignmentStatsPage assignment={assignment} />
-    </Suspense>
-  );
-}
+          <CallPage callAssId={callAssId}>{children}</CallPage>
+        </Suspense>
+      </HomeThemeProvider>
+    );
+  } else {
+    const callUrl = process.env.ZETKIN_GEN2_CALL_URL;
+    const assignmentUrl = `${callUrl}/assignments/${params.callAssId}/call`;
+    redirect(assignmentUrl);
+  }
+};
+
+export default CallPageLayout;
